archive/requirements.txt
neo4j
flask
python-dotenv
PyPDF2
sentence_transformers
faiss-cpu
streamlit
langchain
python-dotenv
langchain_community
chromadb
pysqlite3-binary
pypdf
llama_index
llama-index-llms-ollama
llama-index-embeddings-huggingface
langchain-chroma
langchain-openai
llama_index.llms.ollama
pymongo